pyproject.toml
[tool.poetry]
name = "agentUniverse"
version = "0.1.0a1"
description = "agentUniverse is a framework for developing applications powered by multi-agent base on large language model."

authors = ["AntGroup <jerry.zzw@antgroup.com>"]
repository = "https://github.com/agentuniverse-ai/agentUniverse"
readme = "README_PYPI.md"

packages = [
    { include = "agentuniverse" },
    { include = "agentuniverse_connector" },
    { include = "agentuniverse_extension" },
    { include = "agentuniverse_product" },
]
include = ["*.yaml"]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.0"
cffi = "^1.15.1"
flask = "^2.3.2"
werkzeug = "^3.0.3"
openai = '1.55.3'
tiktoken = '<1.0.0'
loguru = '0.7.2'
flask_cors = '4.0.0'
SQLAlchemy = '2.0.25'
pydantic = "^2.6.4"
gunicorn = "^22.0.0"
chromadb = ">=0.4.24,<1.0.0"
opentelemetry-api = "^1.30.0"
opentelemetry-sdk = "^1.30.0"
opentelemetry-semantic-conventions = ">=0.51b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.30.0"
opentelemetry-exporter-otlp-proto-http = "^1.30.0"
sphinx = "^7.2.6"
Jinja2 = "^3.1.4"
tqdm = "^4.66.3"
sphinx-rtd-theme = "^2.0.0"
aliyun-log-python-sdk = { version = "*", optional = true}
pymilvus = { version = "^2.4.3", optional = true}
googleapis-common-protos = "^1.63.0"
myst-parser = "^2.0.0"
qianfan = "^0.3.12"
dashscope = "^1.19.1"
anthropic = "^0.26.0"
ollama = '^0.2.1'
numpy = '^1.26.0'
pandas = "^2.2.2"
pyarrow = "^16.1.0"
duckduckgo-search = "^6.3.5"
primp = "^0.6.5"
wikipedia= "^1.4.0"
openpyxl = "^3.1.5"
pillow = "^10.4.0"
jieba = "^0.42.1"
networkx = "^3.3"
httpx = ">=0.27.2"
tomli = "^2.2"
mcp = "<1.22.0"
opentracing = ">=2.4.0,<3.0.0"
jsonlines = "^4.0.0"
EbookLib = "^0.18"
beautifulsoup4 = "^4.12.0"
# qdrant-client = "^1.15.1" unsupportable version due to numpy version limit

[tool.poetry.extras]
log_ext = ["aliyun-log-python-sdk"]
store_ext = ["pymilvus"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.6.4"
pre-commit = "^2.20.0"


[[tool.poetry.source]]
name = "china"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
preview = true

[tool.mypy]
files = ["agentuniverse"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["agentuniverse"]